requirements.txt
pyserial-asyncio
smspdudecoder
pydantic
pytest
pytest-asyncio